requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
sqlalchemy>=1.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.5
qrcode[pil]>=7.0.0
reportlab>=3.6.0
jinja2>=3.0.0
python-dotenv>=0.19.0
streamlit>=1.20.0
requests>=2.25.0
pillow>=8.0.0
pyzbar>=0.1.8
opencv-python-headless>=4.5.0
email-validator>=1.1.0
